add files for fine-tuned segmentation with segmenter

+++ segmenter/segm/metrics.py
@@ -107,7 +107,7 @@ def compute_metrics(
         ret_metrics = [ret_metrics["aAcc"], ret_metrics["Acc"], ret_metrics["IoU"]]
         ret_metrics_mean = torch.tensor(
             [
-                np.round(np.nanmean(ret_metric.astype(np.float)) * 100, 2)
+                np.round(np.nanmean(ret_metric.astype(np.float32)) * 100, 2)
                 for ret_metric in ret_metrics
             ],
             dtype=float,
